refactor: move all enums into their own library, and expose them in the API
Currently, this places all enums into a single `enums/index.ts` file. However, this structure lets us easily break things up later on into more files, without needing to change any code which imports `enums/index.ts`.
