+++ src/index.ts
@@ -1,2 +1,7 @@
 export * from './atem'
 export * from './lib/atemState'
+
+import * as Enums from './enums'
+export {
+	Enums
+}
