+++ src/lib/atemState.ts
@@ -1,4 +1,4 @@
-import { Model, TransitionStyle } from '../atem'
+import { Model, TransitionStyle, DVEEffect } from '../enums'
 
 export class AtemState {
 	info = new DeviceInfo()
@@ -152,46 +152,4 @@ export class MediaState {
 	stillPool = {}
 	clipPool = {}
 	players: Array<{ playing: boolean, loop: boolean, atBeginning: boolean, clipFrame: number }> = []
-}
-
-export enum DVEEffect {
-	SwooshTopLeft= 0,
-	SwooshTop = 1,
-	SwooshTopRight = 2,
-	SwooshLeft = 3,
-	SwooshRight = 4,
-	SwooshBottomLeft = 5,
-	SwooshBottom = 6,
-	SwooshBottomRight = 7,
-
-	SpinCCWTopRight = 13,
-	SpinCWTopLeft = 8,
-	SpinCCWBottomRight = 15,
-	SpinCWBottomLeft = 10,
-	SpinCWTopRight = 9,
-	SpinCCWTopLeft = 12,
-	SpinCWBottomRight = 11,
-	SpinCCWBottomLeft = 14,
-
-	SqueezeTopLeft = 16,
-	SqueezeTop = 17,
-	SqueezeTopRight = 18,
-	SqueezeLeft = 19,
-	SqueezeRight = 20,
-	SqueezeBottomLeft = 21,
-	SqueezeBottom = 22,
-	SqueezeBottomRight = 23,
-
-	PushTopLeft = 24,
-	PushTop = 25,
-	PushTopRight = 26,
-	PushLeft = 27,
-	PushRight = 28,
-	PushBottomLeft = 29,
-	PushBottom = 30,
-	PushBottomRight = 31,
-
-	GraphicCWSpin = 32,
-	GraphicCCWSpin = 33,
-	GraphicLogoWipe = 34
 }
